A major refactor to use fewer temporary variables.
This re-works the translation to evaluate the stack
via direct JS expressions rather than a bunch of stack
variables, where possible.  It produces much better
code, at the cost of having to build up an in-memory
representation of each function and the expressions/statements
therein.  I think it'll be a net win overall.